+++ TombEngine/Scripting/Internal/TEN/Flow/Level/FlowLevel.h
@@ -38,7 +38,7 @@ struct Level : public ScriptInterfaceLevel
 	int LevelFarView{ 0 };
 	bool UnlimitedAir{ false };
 	std::vector<InventoryItem> InventoryObjects;
-	short LevelSecrets{ 0 };
+	int LevelSecrets{ 0 };
 
 	RGBAColor8Byte GetFogColor() const override;
 	bool GetFogEnabled() const override;
@@ -57,5 +57,5 @@ struct Level : public ScriptInterfaceLevel
 	short GetFogMaxDistance() const override;
 	short GetFarView() const override;
 	void SetSecrets(short secrets);
-	short GetSecrets() const override;
+	int GetSecrets() const override;
 };
